Extract NotificationDetailHostingController base
Pulls the shared navigation chrome out of the two per-section hosting
controllers — Save bar button (auto-enabled via observation of
`hasUnsavedChanges`), saving spinner driven by `viewModel.isSaving`,
custom back chevron and edge-swipe routed through the discard-changes
action sheet — into a generic `NotificationDetailHostingController<Content>`
base. Each kind's host controller becomes ~15 lines: pass the rooted
detail view in and a scoped `onDiscard` closure pointing at the
relevant `discardStore*Edits()` method on the shared view model.

Drops the duplicated Localization enums; the Save / Back strings now
live on the base and use a kind-agnostic reverse-DNS key. Stock's
detail screen, when it lands, will subclass the same base without
further plumbing.
